+++ petals/src/api/__tests__/util.test.ts
@@ -1,4 +1,4 @@
-import * as util from './util';
+import * as util from '../util';
 
 describe('convertGetQueries', () => {
   it('correct', () => {
